main.go cleaned up, DI registration functions moved to appropriate packages
